Relocate plan write helpers to actor_loop_flow (Issue #681 acceptance)
Move 6 fns + 2 structs (~160 LOC of bodies) covering the "plan
write" cluster out of turn.rs into actor_loop_flow.rs. After this
PR, turn.rs drops below the Issue #681 acceptance criterion
(<= 35,000 LOC).

Moved (turn.rs -> actor_loop_flow.rs):

  Structs:
    - PlanWriteSummary
    - PlanWriteSectionDelta (+ all 5 fields, already pub(super))

  Pure helpers:
    - plan_path_matches
    - plan_section_excerpt
    - plan_write_previous_contents
    - build_plan_write_section_delta
    - plan_write_status
    - plan_phase_from_sections

Adjustments:

  * actor_loop_flow.rs: dropped 7 super::turn:: names from the
    use block (now local); added `use crate::safety::path_guard::
    resolve_user_path;` and `use crate::tools::registry::
    resolve_plan_mode_write_target;` (needed by the moved
    plan_phase_from_sections / build_plan_write_section_delta);
    kept plan_section_body_for_progress + plan_sections_with_content
    as super::turn:: back-imports (still owned by turn.rs).

  * turn.rs: removed now-unused
    `use crate::tools::registry::resolve_plan_mode_write_target`;
    upgraded plan_section_body_for_progress and
    plan_sections_with_content to pub(super) so actor_loop_flow can
    import them; rewrote 4 plan_path_matches and 1
    plan_phase_from_sections call sites to use
    super::actor_loop_flow:: qualification; updated 2 mod tests
    plan_write_status callers to super::super::actor_loop_flow::.

Metrics:

  turn.rs:            35,120 -> 34,958 (-162)
  actor_loop_flow.rs:  4,539 ->  4,708 (+169)

Issue #681 acceptance criterion (turn.rs <= 35,000): MET.
