+++ src/agent/loop_run/actor_loop_flow.rs
@@ -45,10 +45,12 @@ use crate::model_capabilities::model_capabilities;
 use crate::modes::plan_act::{ExecutionMode, PlanStage};
 use crate::ollama::client::AssistantReply;
 use crate::ollama::xml_fallback::ToolCall;
+use crate::safety::path_guard::resolve_user_path;
 use crate::session::feedback::{
     FeedbackFrame, FeedbackFrameDraft, FeedbackKind, build_feedback_frame,
 };
 use crate::session::store::ConversationMessage;
+use crate::tools::registry::resolve_plan_mode_write_target;
 
 use super::Agent;
 use super::active_job_arbiter::{LoopControlAction, RecoveryDispatchGate, RecoveryOwner};
@@ -67,9 +69,8 @@ use super::tool_history::is_plan_file_tool_call;
 use super::tool_policy::EffectiveToolPolicy;
 use super::turn::{
     LOG_ARGS_MAX_CHARS, PLAN_REPEATED_EXPLORATION_BLOCK_THRESHOLD, PlanExplorationKey,
-    PlanWriteSummary, build_plan_write_section_delta, join_sections_for_progress,
-    normalize_exploration_path, plan_path_matches, plan_phase_from_sections, plan_section_excerpt,
-    plan_write_previous_contents, plan_write_status, tool_display, write_stdout_rendered,
+    join_sections_for_progress, normalize_exploration_path, plan_section_body_for_progress,
+    plan_sections_with_content, tool_display, write_stdout_rendered,
 };
 use crate::agent::prompting;
 use crate::session::compact::approximate_token_count;
@@ -4536,4 +4537,172 @@ pub(super) fn should_apply_repo_change_quality_gate(
     mode == ExecutionMode::Act
         && (action_expectation == recovery::ActionExpectation::RepoChange
             || active_task_expects_repo_change)
+}
+
+pub(super) struct PlanWriteSummary {
+    pub(super) action: String,
+    pub(super) note: Option<String>,
+    pub(super) status: Option<String>,
+    pub(super) phase: String,
+    pub(super) signature: String,
+}
+
+pub(super) fn plan_path_matches(
+    raw_path: &str,
+    work_root: &Path,
+    plan_path: Option<&Path>,
+) -> bool {
+    resolve_plan_mode_write_target(work_root, raw_path, plan_path)
+        .ok()
+        .flatten()
+        .is_some()
+}
+
+pub(super) fn plan_section_excerpt(contents: &str, sections: &[&str]) -> Option<String> {
+    for section in sections {
+        let Some(body) = plan_section_body_for_progress(contents, section) else {
+            continue;
+        };
+        for line in body.lines().map(str::trim) {
+            if line.is_empty()
+                || line == "-"
+                || matches!(
+                    line,
+                    "1." | "2."
+                        | "3."
+                        | "1. First slice:"
+                        | "2. Next phases:"
+                        | "3. Review checkpoint:"
+                )
+            {
+                continue;
+            }
+            let cleaned = line.trim_start_matches("- ").trim();
+            return Some(format!(
+                "{section}: {}",
+                truncate(&sanitize_for_progress(cleaned), 72)
+            ));
+        }
+    }
+    None
+}
+
+pub(super) fn plan_write_previous_contents(
+    raw_path: &str,
+    work_root: &Path,
+    plan_path: Option<&Path>,
+) -> String {
+    if raw_path.is_empty() {
+        String::new()
+    } else if plan_path_matches(raw_path, work_root, plan_path) {
+        plan_path
+            .and_then(|path| std::fs::read_to_string(path).ok())
+            .unwrap_or_default()
+    } else {
+        resolve_user_path(work_root, raw_path)
+            .ok()
+            .and_then(|path| std::fs::read_to_string(path).ok())
+            .unwrap_or_default()
+    }
+}
+
+pub(super) struct PlanWriteSectionDelta {
+    pub(super) previous_sections: Vec<&'static str>,
+    pub(super) current_sections: Vec<&'static str>,
+    pub(super) added_sections: Vec<&'static str>,
+    pub(super) removed_sections: Vec<&'static str>,
+    pub(super) focus_sections: Vec<&'static str>,
+}
+
+pub(super) fn build_plan_write_section_delta(
+    previous: &str,
+    new_text: &str,
+) -> PlanWriteSectionDelta {
+    let previous_sections = plan_sections_with_content(previous);
+    let current_sections = plan_sections_with_content(new_text);
+    let changed_sections = current_sections
+        .iter()
+        .copied()
+        .filter(|section| {
+            let old_body = plan_section_body_for_progress(previous, section).unwrap_or_default();
+            let new_body = plan_section_body_for_progress(new_text, section).unwrap_or_default();
+            sanitize_for_progress(old_body) != sanitize_for_progress(new_body)
+        })
+        .collect::<Vec<_>>();
+    let added_sections = current_sections
+        .iter()
+        .copied()
+        .filter(|section| !previous_sections.contains(section))
+        .collect::<Vec<_>>();
+    let removed_sections = previous_sections
+        .iter()
+        .copied()
+        .filter(|section| !current_sections.contains(section))
+        .collect::<Vec<_>>();
+    let focus_sections = if !changed_sections.is_empty() {
+        changed_sections
+    } else if !current_sections.is_empty() {
+        current_sections.clone()
+    } else {
+        Vec::new()
+    };
+    PlanWriteSectionDelta {
+        previous_sections,
+        current_sections,
+        added_sections,
+        removed_sections,
+        focus_sections,
+    }
+}
+
+pub(super) fn plan_write_status(new_text: &str, delta: isize) -> Option<String> {
+    let next = lifecycle::plan_next_stage_sections(new_text);
+    if lifecycle::plan_missing_sections(new_text).is_empty() {
+        Some(format!("Approval ready | delta {delta:+}B"))
+    } else if !next.is_empty() {
+        Some(format!(
+            "Next: {} | delta {delta:+}B",
+            join_sections_for_progress(&next)
+        ))
+    } else {
+        Some(format!("delta {delta:+}B"))
+    }
+}
+
+pub(super) fn plan_phase_from_sections(
+    sections: &[&str],
+    current_stage: PlanStage,
+    approval_ready: bool,
+) -> &'static str {
+    if approval_ready {
+        "Approval review"
+    } else if sections.iter().any(|section| {
+        matches!(
+            *section,
+            "First Action"
+                | "Verification"
+                | "Execution Plan"
+                | "Verification Plan"
+                | "Risks / Fallbacks"
+        )
+    }) {
+        "Define next action"
+    } else if sections
+        .iter()
+        .any(|section| matches!(*section, "Acceptance Criteria" | "Quality Bar"))
+    {
+        "Define quality bar"
+    } else if sections
+        .iter()
+        .any(|section| matches!(*section, "Goal" | "Constraints" | "Deliverables"))
+    {
+        "Draft foundation"
+    } else {
+        match current_stage {
+            PlanStage::Stage1 => "Draft foundation",
+            PlanStage::Stage2 => "Define next action",
+            PlanStage::Stage3 => "Approval review",
+            PlanStage::Ready => "Approval review",
+        }
+    }
 }
